apps/python/requirements.txt
fastapi[all]==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
pydantic==2.5.0
python-multipart==0.0.6
boto3==1.34.0
python-dotenv==1.0.0
PyPDF2==3.0.1
fitz==0.0.1.dev2
nltk==3.8.1
spacy==3.7.2
elevenlabs==0.2.26
openai==1.3.7
requests==2.31.0
aiofiles==23.2.0
redis==5.0.1
asyncio-mqtt==0.16.1
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1
pydub==0.25.1